openzim/requirements.txt
libzim>=2.1.0
docopt>=0.6.2
path.py==12.5.0
python-magic==0.4.27
requests>=2.28.2
frontmatter>=1.0.0